requirements.txt
# Minimal requirements
Django>=1.7,<1.8
wagtail==0.8.4

# Recommended components (require additional setup):
# psycopg2==2.5.2
# elasticsearch==1.1.1

# Recommended components to improve performance in production:
# django-redis-cache==0.13.0
# django-celery==3.1.10

djangorestframework>=3.0
